유틸리티 함수 파일 분할
